third_party/WebKit/LayoutTests/fast/harness/perftests/runs-per-second-iterations-expected.txt

This test verifies PerfTestRunner.runPerSecond() calls runFunction as many times as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Returning times: [0, 10, 30, 60, 200, 200, 200, 200, 750]
PASS callsInIterations[0] is 1
PASS callsInIterations[1] is 10
PASS callsInIterations[1] + 5 < callsInIterations[2] is true
PASS callsInIterations[2] + 10 < callsInIterations[3] is true
PASS callsInIterations[3] is callsInIterations[4]
PASS callsInIterations[4] is callsInIterations[5]
PASS callsInIterations[5] is callsInIterations[6]
PASS callsInIterations[6] is callsInIterations[7]
PASS callsInIterations[7] is callsInIterations[8]
PASS callsInIterations[9] is undefined.
